requirements.txt
aiohttp~=3.9.5
aiosignal~=1.3.1
annotated-types~=0.7.0
anyio~=4.4.0
appdirs~=1.4.4
attrs~=23.2.0
certifi~=2024.2.2
cffi~=1.16.0
charset-normalizer~=3.3.2
click~=8.1.7
colorama~=0.4.6
cryptography~=42.0.8
dnspython~=2.6.1
email_validator~=2.1.1
fastapi~=0.111.0
fastapi-cli~=0.0.4
frozenlist~=1.4.1
h11~=0.14.0
httpcore~=1.0.5
httptools~=0.6.1
httpx~=0.27.0
idna~=3.7
importlib_metadata~=7.1.0
Jinja2~=3.1.4
markdown-it-py~=3.0.0
MarkupSafe~=2.1.5
mdurl~=0.1.2
multidict~=6.0.5
orjson~=3.10.5
pdfminer.six~=20231228
pdfplumber~=0.11.2
pillow~=10.4.0
pycparser~=2.22
pydantic~=2.7.4
pydantic_core~=2.18.4
pyee~=11.1.0
Pygments~=2.18.0
PyMuPDF~=1.24.7
PyMuPDFb~=1.24.6
pypdf~=4.3.0
pypdfium2~=4.30.0
pyppeteer~=2.0.0
python-dotenv~=1.0.1
python-multipart~=0.0.9
PyYAML~=6.0.1
requests~=2.32.3
rich~=13.7.1
setuptools~=58.1.0
shellingham~=1.5.4
sniffio~=1.3.1
soupsieve~=2.5
starlette~=0.37.2
tqdm~=4.66.4
typer~=0.12.3
typing_extensions~=4.12.0
ujson~=5.10.0
urllib3~=1.26.18
uvicorn~=0.30.1
watchfiles~=0.22.0
websockets~=10.4
yarl~=1.9.4
zipp~=3.19.0
